Refreactor pairs func. into two seperate function exposed to python.
The pairs function was unclear and nested the logic of energy and spin
summation unneccesarily tightly together. To increase simplicitly they
are now separeted.

The decision to expose the function as part of PyTandem python was
made in the spirit of moving more of the wave function logic (as that
is O(n)) to python while keeping the core tandem logic (that is
O(n^2)) completley in c++ for speed.

The pairs function was initially written to be a part of
tandem_function_interface functionality, otherwise I would have
probably implemented it in python directly.