Add Hartre fock logic test.
Tests the core logic (not c++ implementation) of how the hartree fock
D matrices are constructed into wave functions.

This test was born out of depsperation, trying to debugg the hartree
fock c++ implementation but going nowhere. turned out the logic was
fine. A factor -1/4 in the interaction scaling was to blame...